dash-tui: common db detection with metrics
